+++ src/components/projects.tsx
@@ -3,7 +3,7 @@ import { getScopedI18n } from "@/locales/server";
 import { projects } from "../data/project-data";
 import { ArrowRightIcon } from "@phosphor-icons/react/dist/ssr";
 import { TransitionLink } from "./utils/transition-link";
-import Button from "./ui/button";
+import Button from "./ui/button/button";
 
 export default async function Projects() {
   const scopedT = await getScopedI18n("frontPage.projects");
