+++ src/components/ui/button/button.module.css
@@ -1,0 +1,102 @@
+.root {
+  font-family: var(--font-heading);
+  display: inline-flex;
+  cursor: pointer;
+  align-items: center;
+  justify-content: center;
+  gap: 0.5rem;
+  border-radius: 9999px;
+  font-size: 0.875rem;
+  font-weight: 600;
+  white-space: nowrap;
+  transition-property:
+    color, background-color, border-color, text-decoration-color, fill, stroke;
+  transition-duration: 0.15s;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  border: none;
+  background: transparent;
+
+  & svg {
+    pointer-events: none;
+    width: 1rem;
+    height: 1rem;
+    flex-shrink: 0;
+  }
+
+  &:disabled {
+    pointer-events: none;
+    opacity: 0.5;
+  }
+
+  &:not(:hover)::before {
+    transition-duration: 200ms;
+  }
+
+  @media (min-width: 768px) {
+    font-size: 1rem;
+  }
+
+  @media (min-width: 1280px) {
+    font-size: 1.125rem;
+  }
+}
+
+.size-default {
+  height: 2.25rem;
+  padding: 0.5rem 1rem;
+}
+
+.size-sm {
+  height: 2rem;
+  padding: 0 0.75rem;
+  font-size: 0.75rem;
+}
+
+.size-lg {
+  height: 2.5rem;
+  padding: 0 2rem;
+}
+
+.size-icon {
+  height: 2rem;
+  width: 2rem;
+}
+
+.variant-default {
+  position: relative;
+  background-color: var(--whitebg);
+  color: #2c2c2c;
+  filter: drop-shadow(0 1px 2px rgba(0, 0, 0, 0.05));
+  transition-property: all;
+  transition-duration: 0.15s;
+}
+
+.variant-slider {
+  background-color: var(--whitebg);
+  color: var(--background);
+
+  @media (min-width: 1024px) {
+    transition-property: all;
+    transition-duration: 0.15s;
+
+    &:hover {
+      transform: scale(1.1);
+    }
+  }
+
+  @media (prefers-color-scheme: dark) {
+    background-color: #2c2c2c;
+  }
+}
+
+:global(.dark) .variant-slider {
+  background-color: #2c2c2c;
+}
+
+.content {
+  position: relative;
+  z-index: 10;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
